Refactored BidirPathTracer a little. Fixed non-post pro aperture diffraction for bidir (BidirPathTracer). Added sun_glare_test_indirect.igs.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@4070 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ simpleraytracer/camera.h
@@ -119,7 +119,7 @@ public:
 	virtual Real getExposureDuration() const = 0;
 
 
-	virtual const Vec3Type diffractRay(const SamplePair& samples, const Vec3Type& dir, const SpectralVector& wavelengths, double direction_sign, double time, SpectralVector& weights_out) const = 0;
+	virtual const Vec3Type diffractRay(const SamplePair& samples, const Vec3Type& dir, const SpectralVector& wavelengths, float direction_sign, double time, SpectralVector& weights_out) const = 0;
 
 
 	virtual void prepareForDiffractionFilter(int main_buffer_width, int main_buffer_height, int ssf_) = 0;
